70. 모임 참가 신청 수락 취소 및 출석 체크

+++ src/main/java/com/mystudy/event/EventService.java
@@ -50,10 +50,28 @@ public class EventService {
 		}
 	}
 
-	public void cancelEnrollment(Event event, Member account) {
-		Enrollment enrollment = enrollmentRepository.findByEventAndMember(event, account);
-		event.removeEnrollment(enrollment);
-		enrollmentRepository.delete(enrollment);
-		event.acceptNextWaitingEnrollment();
+	public void cancelEnrollment(Event event, Member member) {
+		Enrollment enrollment = enrollmentRepository.findByEventAndMember(event, member);
+		if (!enrollment.isAttended()) {
+			event.removeEnrollment(enrollment);
+			enrollmentRepository.delete(enrollment);
+			event.acceptNextWaitingEnrollment();
+		}
+	}
+
+	public void acceptEnrollment(Event event, Enrollment enrollment) {
+		event.accept(enrollment);
+	}
+
+	public void rejectEnrollment(Event event, Enrollment enrollment) {
+		event.reject(enrollment);
+	}
+
+	public void checkInEnrollment(Enrollment enrollment) {
+		enrollment.setAttended(true);
+	}
+
+	public void cancelCheckInEnrollment(Enrollment enrollment) {
+		enrollment.setAttended(false);
 	}
 }
